services/backend/requirements/base.txt
alabaster==0.7.16
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==23.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.10.0
astroid==3.0.3
asttokens==3.0.0
attrs==25.4.0
Authlib==1.6.5
babel==2.17.0
backcall==0.2.0
billiard==4.2.2
bleach==6.3.0
blinker==1.9.0
boto3==1.34.0
botocore==1.34.162
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coverage==7.11.0
cron_descriptor==2.0.6
cryptography==46.0.3
decorator==5.2.1
dill==0.4.0
dj-database-url==2.1.0
Django>=5.1.14,<5.2.0  # Security fix for CVE-2025-64459 (SQL injection) and GHSA-frmv-pr5f-9mcr
django-celery-beat==2.8.1
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-opensearch-dsl>=0.5.0,<1.0.0
django-email-validation==2.1.1
django-extensions==3.2.3
django-filter==24.2
django-prometheus==2.3.1
django-ratelimit==4.1.0
django-redis==5.4.0
django-storages==1.14.2
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
docutils==0.20.1
dparse==0.6.4
drf-spectacular==0.27.0
opensearch-dsl>=2.1.0,<3.0.0
opensearch-py>=2.5.0,<3.0.0
executing==2.2.1
factory-boy==3.3.0
Faker==20.1.0
filelock>=3.20.1,<4.0.0  # Security fix for CVE-2025-68146 (TOCTOU race condition)
flower==2.0.1
gprof2dot==2025.4.14
gunicorn==23.0.0
h11==0.16.0
hiredis==3.3.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.14.0
idna==3.11
imagesize==1.4.1
inflection==0.5.1
iniconfig==2.3.0
ipdb==0.13.13
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow>=4.1.2,<5.0.0  # Update to latest secure version
matplotlib-inline==0.2.1
mccabe==0.7.0
mdurl==0.1.2
packaging==25.0
parso==0.8.5
pathspec==0.12.1
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.5.0
pluggy==1.6.0
prometheus_client==0.23.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.11
pure_eval==0.2.3
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pyflakes>=3.4.0,<3.5.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==7.4.3
pytest-cov==4.1.0
pytest-django==4.7.0
python-crontab==3.3.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.3
redis==5.0.1
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rpds-py==0.28.0
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.14
s3transfer==0.9.0
safety==3.7.0
safety-schemas==0.0.16
sentry-sdk==2.14.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
snowballstemmer==3.0.1
Sphinx==7.2.6
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.3
stack-data==0.6.3
stevedore==5.5.0
stripe==7.10.0
tenacity==9.1.2
tomlkit==0.13.3
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3>=2.5.2,<3.0.0  # Security fix for CVE-2025-66471 and CVE-2025-66418
vine==5.1.0
wcwidth==0.2.14
webencodings==0.5.1
wheel==0.45.1
whitenoise==6.6.0

# OpenTelemetry for Distributed Tracing
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation==0.42b0
opentelemetry-instrumentation-django==0.42b0
opentelemetry-instrumentation-psycopg2==0.42b0
opentelemetry-instrumentation-redis==0.42b0
opentelemetry-instrumentation-requests==0.42b0
opentelemetry-instrumentation-celery==0.42b0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-exporter-otlp==1.21.0
